+++ changelog.md
@@ -170,3 +170,7 @@
 
 - keep entire treeview visible
 
+-  clear txtNewBen after insertion of new ben into lbBenDest
+
+-  clear txtNewAttr after insertion of new attr into lbAttrDest
+